refactor physical layout
