+++ controller.test.ts
@@ -3,9 +3,9 @@
  */
 import {describe, expect, test} from '@jest/globals';
 const cloneDeep = require('lodash.clonedeep');
+import BoardState from './model';
 import
 {
-    BoardState,
     MoveCard,
     MoveList,
     AddCard,
@@ -17,7 +17,7 @@ import
     ChangeCardLabels,
     ChangeCardNotes,
     fillerStr
-} from './script';
+} from './controller';
 
 /*
 * The object is wrapped in a function as a simple
